25/pyproject.toml
[project]
name = "sd-25"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datasets>=4.0.0",
    "dspy>=3.0.1",
    "mlflow>=3.2.0",
    "openai==1.99.5",
    "python-dotenv>=1.1.1",
]
